+++ unitylibs/core/workflow/workflow-firefly/action-binder.js
@@ -227,15 +227,15 @@ export default class ActionBinder {
   }
 
   /**
-   * When Generate runs from the full style launcher, the connector payload should include the selected style name
+   * When Generate runs from the prompt-with-style-select UI, the connector payload should include the selected style name
    * at the end of the prompt (DOM label under the thumbnail).
    *
    * @param {string} query
    * @returns {string}
    */
-  appendSelectedStyleNameForStyleLauncher(query) {
+  appendSelectedStyleNameForPromptWithStyleSelect(query) {
     const root = this.block;
-    if (!root?.classList?.contains('unity-style-launcher-full')) return query;
+    if (!root?.classList?.contains('unity-prompt-with-style-select')) return query;
     const styleName = root.querySelector('.unity-slf-style-item.selected .unity-slf-style-label')?.textContent?.trim();
     if (!styleName || !query) return query;
     return `${query} ${styleName}`;
@@ -276,7 +276,7 @@ export default class ActionBinder {
     } else {
       this.query = this.inputField.value.trim();
     }
-    this.query = this.appendSelectedStyleNameForStyleLauncher(this.query);
+    this.query = this.appendSelectedStyleNameForPromptWithStyleSelect(this.query);
     const selectedVerbType = `text-to-${currentVerb}`;
     const action = (this.id || !!override ? 'prompt-suggestion' : 'generate');
     const eventData = { assetId: this.id, verb: selectedVerbType, action };
