+++ unitylibs/core/workflow/workflow-firefly/widget.js
@@ -1,7 +1,11 @@
 /* eslint-disable class-methods-use-this */
 
-import { createTag, getConfig, getUnityLibs, unityConfig } from '../../../scripts/utils.js';
+import { createTag, getUnityLibs } from '../../../scripts/utils.js';
 
+/**
+ * Shared Firefly widget: model picker, Generate button, dropdown plumbing used by
+ * prompt-with-style-select and the hero prompt bar ({@link PromptWidget} adds verb/prompt-dropdown UI).
+ */
 export default class UnityWidget {
   constructor(target, el, workflowCfg, spriteCon) {
     this.el = el;
@@ -25,64 +29,28 @@ export default class UnityWidget {
     this.lanaOptions = { sampleRate: 100, tags: 'Unity-FF' };
     this.sound = { audio: null, currentTile: null, currentUrl: '' };
     this.durationCache = new Map();
-    /** @type {HTMLElement | null} Set when full style launcher mounts as a sibling of the Unity block */
-    this.styleLauncherRoot = null;
+    /** @type {HTMLElement | null} Set when prompt-with-style-select UI mounts as a sibling of the Unity block */
+    this.promptWithStyleSelectRoot = null;
   }
 
+  /**
+   * Delegates to `core/widgets/*` entry points.
+   * - `mountInUnityBlock` → prompt-with-style-select
+   * - else → marquee prompt widget (subclass with verb + prompt suggestions)
+   */
   async initWidget() {
+    const widgetsBase = `${getUnityLibs()}/core/widgets`;
     if (this.workflowCfg.targetCfg?.mountInUnityBlock) {
-      // Resolve via getUnityLibs() so the request matches other workflow modules (relative ./ URLs can 404 under import maps / CDN).
-      const styleLauncherUrl = `${getUnityLibs()}/core/widgets/style-launcher/style-launcher-full.js`;
-      const { parseStyleLauncherAuthoring, mountStyleLauncherFullUI } = await import(styleLauncherUrl);
-      const parsed = parseStyleLauncherAuthoring(this.el);
-      if (!parsed.styles.length) return this.workflowCfg.targetCfg.actionMap;
-      await mountStyleLauncherFullUI(this, parsed);
-      return this.workflowCfg.targetCfg.actionMap;
+      const { initPromptWithStyleSelectWidget } = await import(`${widgetsBase}/prompt-with-style-select/prompt-with-style-select.js`);
+      return initPromptWithStyleSelectWidget(this);
     }
-    const [widgetWrap, widget, unitySprite] = ['ex-unity-wrap', 'ex-unity-widget', 'unity-sprite-container']
-      .map((c) => createTag('div', { class: c }));
-    this.widgetWrap = widgetWrap;
-    this.widget = widget;
-    unitySprite.innerHTML = this.spriteCon;
-    this.widgetWrap.append(unitySprite);
-    this.workflowCfg.placeholder = this.popPlaceholders();
-    const hasPromptPlaceholder = !!this.el.querySelector('.icon-placeholder-prompt');
-    const hasSuggestionsPlaceholder = !!this.el.querySelector('.icon-placeholder-suggestions');
-    const hasModels = !!this.el.querySelector('[class*="icon-model"]');
-    this.hasModelOptions = hasModels;
-    this.hasPromptSuggestions = hasPromptPlaceholder && hasSuggestionsPlaceholder;
-    if (this.hasModelOptions) await this.getModel();
-    const inputWrapper = this.createInpWrap(this.workflowCfg.placeholder);
-    await this.ensureSoundModuleLoaded();
-    let dropdown = null;
-    if (this.hasPromptSuggestions) dropdown = await this.genDropdown(this.workflowCfg.placeholder);
-    const comboboxContainer = createTag('div', { class: 'autocomplete' });
-    comboboxContainer.append(inputWrapper);
-    if (dropdown) comboboxContainer.append(dropdown);
-    this.widget.append(comboboxContainer);
-    this.addWidget();
-    if (this.workflowCfg.targetCfg.floatPrompt) this.initIO();
-    return this.workflowCfg.targetCfg.actionMap;
+    const { initPromptWidget } = await import(`${widgetsBase}/prompt-widget/prompt-widget.js`);
+    return initPromptWidget(this);
   }
 
+  /** No-op on base; marquee subclass loads sound augmentation for the `sound` verb. */
   async ensureSoundModuleLoaded() {
-    if (this.selectedVerbType !== 'sound' || this.soundAugmented) return;
-    try {
-      const { default: augmentSound } = await import('./sound-utils.js');
-      augmentSound(this);
-      this.soundAugmented = true;
-    } catch (e) {
-      window.lana?.log(`Message: Error loading sound module, Error: ${e}`, this.lanaOptions);
-    }
-  }
-
-  popPlaceholders() {
-    return Object.fromEntries(
-      [...this.el.querySelectorAll('[class*="placeholder"]')].map((element) => [
-        element.classList[1]?.replace('icon-', '') || '',
-        element.closest('li')?.innerText || '',
-      ]).filter(([key]) => key),
-    );
+    await Promise.resolve();
   }
 
   showVerbMenu(selectedElement) {
@@ -262,67 +230,6 @@ export default class UnityWidget {
     });
   }
 
-  verbDropdown() {
-    const verbs = this.el.querySelectorAll('[class*="icon-verb"]');
-    const inputPlaceHolder = this.el.querySelector('.icon-placeholder-input').parentElement.textContent;
-    const selectedVerbType = verbs[0]?.className.split('-')[2];
-    const selectedVerb = verbs[0]?.nextElementSibling;
-    const selectedElement = createTag('button', {
-      class: 'selected-verb',
-      'aria-expanded': 'false',
-      'aria-controls': 'media-menu',
-      'aria-label': 'media type',
-      'aria-haspopup': 'listbox',
-      role: 'combobox',
-      'aria-labelledby': 'listbox-label',
-      'data-selected-verb': selectedVerbType,
-    }, `${selectedVerb?.textContent.trim()}`);
-    this.selectedVerbType = selectedVerbType;
-    this.widgetWrap.setAttribute('data-selected-verb', this.selectedVerbType);
-    this.selectedVerbText = selectedVerb?.textContent.trim();
-    if (verbs.length <= 1) {
-      selectedElement.setAttribute('disabled', 'true');
-      return [selectedElement];
-    }
-    this.widgetWrap.classList.add('verb-options');
-    const menuIcon = createTag('span', { class: 'menu-icon' }, '<svg><use xlink:href="#unity-chevron-icon"></use></svg>');
-    const verbList = createTag('ul', { class: 'verb-list', id: 'media-menu', role: 'listbox', 'aria-labelledby': 'listbox-label' });
-    verbList.setAttribute('style', 'display: none;');
-    selectedElement.append(menuIcon);
-    const handleDocumentClick = (e) => {
-      const menuContainer = selectedElement.parentElement;
-      if (!menuContainer.contains(e.target)) {
-        document.removeEventListener('click', handleDocumentClick);
-        menuContainer.classList.remove('show-menu');
-        selectedElement.setAttribute('aria-expanded', 'false');
-      }
-    };
-    selectedElement.addEventListener('click', (e) => {
-      e.stopPropagation();
-      this.hidePromptDropdown(selectedElement);
-      this.showVerbMenu(selectedElement);
-      document.addEventListener('click', handleDocumentClick);
-    }, true);
-    selectedElement.addEventListener('keydown', (e) => {
-      if (e.key === 'Enter' || e.key === ' ') {
-        this.hidePromptDropdown(selectedElement);
-        this.showVerbMenu(selectedElement);
-      }
-      if (e.key === 'Escape' || e.code === 27) {
-        selectedElement.parentElement.classList?.remove('show-menu');
-        selectedElement.focus();
-      }
-    });
-    verbs[0]?.classList.add('selected');
-    const verbsData = Array.from(verbs).map((verb) => ({
-      name: verb.nextElementSibling?.textContent.trim(),
-      type: verb.classList[1].split('-')[2],
-      icon: verb.nextElementSibling?.href,
-    }));
-    this.createDropdownItems(verbsData, verbList, selectedElement, menuIcon, inputPlaceHolder, false);
-    return [selectedElement, verbList];
-  }
-
   modelDropdown() {
     if (!this.hasModelOptions) return [];
     const models = Array.isArray(this.models)
@@ -385,127 +292,6 @@ export default class UnityWidget {
     return [selectedElement, listItems];
   }
 
-  createInpWrap(ph) {
-    const inpWrap = createTag('div', { class: 'inp-wrap' });
-    const actWrap = createTag('div', { class: 'act-wrap' });
-    const inpField = createTag('textarea', {
-      id: 'promptInput',
-      class: 'inp-field',
-      type: 'text',
-      placeholder: ph['placeholder-input'],
-      'aria-autocomplete': 'list',
-      'aria-haspopup': 'listbox',
-      'aria-controls': 'prompt-dropdown',
-      'aria-activedescendant': '',
-    });
-    const dropdown = this.widget.querySelector('.prompt-dropdown-container');
-    inpField.addEventListener('focus', () => this.hidePromptDropdown());
-    inpField.addEventListener('click', () => this.resetAllSoundVariations?.(dropdown));
-    inpField.addEventListener('input', () => this.resetAllSoundVariations?.(dropdown));
-    const verbDropdown = this.verbDropdown();
-    const modelDropdown = this.modelDropdown();
-    const genBtn = this.createActBtn(this.el.querySelector('.icon-generate')?.closest('li'), 'gen-btn');
-    actWrap.append(genBtn);
-    const actionContainer = createTag('div', { class: 'action-container' });
-    const hasPromptLabel = this.el.querySelector('.icon-placeholder-prompt-label');
-    if (hasPromptLabel && ph['placeholder-prompt-label']) {
-      const promptLabel = createTag('label', { for: 'promptInput', class: 'inp-field-label' }, ph['placeholder-prompt-label']);
-      inpWrap.appendChild(promptLabel);
-    }
-    if (verbDropdown.length > 1) {
-      const verbBtn = createTag('div', { class: 'verbs-container', 'aria-label': 'Media options' });
-      verbBtn.append(...verbDropdown);
-      actionContainer.append(verbBtn);
-      inpWrap.append(inpField, actionContainer, actWrap);
-    } else {
-      inpWrap.append(inpField, actWrap);
-    }
-    if (modelDropdown.length > 1) {
-      const modelBtn = createTag('div', { class: 'models-container', 'aria-label': 'Model options' });
-      modelBtn.append(...modelDropdown);
-      actionContainer.append(modelBtn);
-    }
-    return inpWrap;
-  }
-
-  getLimitedDisplayPrompts(prompts) {
-    const shuffled = prompts.sort(() => 0.5 - Math.random());
-    return shuffled.slice(0, 3).map(({ prompt, assetid, variations }) => ({
-      prompt,
-      assetid,
-      variations,
-    }));
-  }
-
-  addPromptItemsToDropdown(dropdown, prompts, placeholder) {
-    this.promptItems = [];
-    prompts.forEach(({ prompt, assetid, variations }, idx) => {
-      const item = createTag('li', {
-        id: assetid,
-        class: 'drop-item',
-        role: 'option',
-        tabindex: '0',
-        'aria-label': prompt,
-        'aria-description': `${placeholder['placeholder-prompt']} ${placeholder['placeholder-suggestions']}`,
-        'daa-ll': `${prompt.slice(0, 20)}--${this.selectedVerbType}--Prompt suggestion`,
-      });
-      const iconWrap = createTag('span', { class: 'prompt-icon' }, '<svg><use xlink:href="#unity-prompt-icon"></use></svg>');
-      const text = createTag('span', { class: 'drop-text' }, prompt);
-      item.append(iconWrap, text);
-      dropdown.append(item);
-      this.promptItems.push(item);
-
-      if (this.selectedVerbType === 'sound') this.addSoundSuggestionHandlers(dropdown, item, { prompt, variations }, idx + 1);
-    });
-  }
-
-  async genDropdown(ph) {
-    if (!this.hasPromptSuggestions) return null;
-    const promptDropdownContainer = createTag('div', { class: 'prompt-dropdown-container drop hidden', 'aria-hidden': 'true' });
-    const dd = createTag('ul', {
-      id: 'prompt-dropdown',
-      class: 'prompt-suggestions-list',
-      'daa-lh': 'Marquee',
-      role: 'listbox',
-      'aria-labelledby': 'promptInput',
-    });
-    const titleCon = createTag('div', { class: 'drop-title-con' });
-    const title = createTag('span', { class: 'drop-title', id: 'prompt-suggestions' }, `${ph['placeholder-prompt']} ${ph['placeholder-suggestions']}`);
-    titleCon.append(title);
-    const prompts = await this.getPrompt(this.selectedVerbType);
-    const limited = this.getLimitedDisplayPrompts(prompts);
-    this.addPromptItemsToDropdown(dd, limited, ph);
-    promptDropdownContainer.append(titleCon, dd, this.createFooter(ph));
-    if (!this.outsideDropdownHandler) {
-      this.outsideDropdownHandler = (ev) => {
-        if (this.widgetWrap && window.getComputedStyle(this.widgetWrap).pointerEvents === 'none') return;
-        const wrapper = this.widget.querySelector('.autocomplete');
-        if (wrapper && wrapper.contains(ev.target)) return;
-        if (promptDropdownContainer && !promptDropdownContainer.classList.contains('hidden') && !promptDropdownContainer.contains(ev.target)) {
-          this.hidePromptDropdown();
-        }
-      };
-      setTimeout(() => document.addEventListener('click', this.outsideDropdownHandler, true), 0);
-    }
-    return promptDropdownContainer;
-  }
-
-  createFooter(ph) {
-    const footer = createTag('div', { class: 'drop-footer' });
-    const tipEl = this.el.querySelector('.icon-tip')?.closest('li');
-    const tipCon = createTag('div', { id: 'tip-content', class: 'tip-con', tabindex: '-1', role: 'note', 'aria-label': `${ph['placeholder-tip']} ${tipEl?.innerText}` }, '<svg><use xlink:href="#unity-info-icon"></use></svg>');
-    const tipText = createTag('span', { class: 'tip-text', id: 'tip-text' }, `${ph['placeholder-tip']}:`);
-    const tipDesc = createTag('span', { class: 'tip-desc', id: 'tip-desc' }, tipEl?.innerText || '');
-    tipCon.append(tipText, tipDesc);
-    const legalEl = this.el.querySelector('.icon-legal')?.closest('li');
-    const legalCon = createTag('div', { class: 'legal-con' });
-    const legalLink = legalEl?.querySelector('a');
-    const legalText = createTag('a', { href: legalLink?.href || '#', class: 'legal-text' }, legalLink?.innerText || 'Legal');
-    legalCon.append(legalText);
-    footer.append(tipCon, legalCon);
-    return footer;
-  }
-
   createActBtn(cfg, cls) {
     if (!cfg) return null;
     const txt = cfg.innerText?.trim();
@@ -516,44 +302,6 @@ export default class UnityWidget {
     if (txt) btn.append(createTag('div', { class: 'btn-txt' }, txt.split('\n')[0]));
     this.genBtn = btn;
     return btn;
-  }
-
-  addWidget() {
-    if (this.workflowCfg.targetCfg?.mountInUnityBlock) return;
-    const interactArea = this.target.querySelector('.copy');
-    const para = interactArea?.querySelector(this.workflowCfg.targetCfg.target);
-    this.widgetWrap.append(this.widget);
-    if (para && this.workflowCfg.targetCfg.insert === 'before') para.before(this.widgetWrap);
-    else if (para) para.after(this.widgetWrap);
-    else interactArea?.appendChild(this.widgetWrap);
-  }
-
-  async loadPrompts() {
-    const { locale } = getConfig();
-    const { origin } = window.location;
-    const baseUrl = (origin.includes('.aem.') || origin.includes('.hlx.'))
-      ? `https://main--unity--adobecom.${origin.includes('.hlx.') ? 'hlx' : 'aem'}.live`
-      : origin;
-    const promptFile = locale.prefix && locale.prefix !== '/'
-      ? `${baseUrl}${locale.prefix}/unity/configs/prompt/firefly-prompt.json`
-      : `${baseUrl}/unity/configs/prompt/firefly-prompt.json`;
-    const promptRes = await fetch(promptFile);
-    if (!promptRes.ok) {
-      throw new Error('Failed to fetch prompts.');
-    }
-    const promptJson = await promptRes.json();
-    this.prompts = this.createPromptMap(promptJson?.content?.data);
-  }
-
-  async getPrompt(verb) {
-    if (!this.hasPromptSuggestions) return [];
-    try {
-      if (!this.prompts || Object.keys(this.prompts).length === 0) await this.loadPrompts();
-      return (this.prompts?.[verb] || []).filter((item) => item.prompt && item.prompt.trim() !== '');
-    } catch (e) {
-      window.lana?.log(`Message: Error loading promts, Error: ${e}`, this.lanaOptions);
-      return [];
-    }
   }
 
   async loadModels() {
@@ -581,39 +329,11 @@ export default class UnityWidget {
     }
   }
 
-  createPromptMap(data) {
-    const promptMap = {};
-    if (Array.isArray(data)) {
-      data.forEach((item) => {
-        const itemEnv = item.env || 'prod';
-        if (item.verb && item.prompt && itemEnv === unityConfig.env) {
-          if (!promptMap[item.verb]) promptMap[item.verb] = [];
-          let variations = [];
-          if (item.verb === 'sound') {
-            const ph = this.workflowCfg?.placeholder || {};
-            const labels = Object.keys(ph)
-              .filter((k) => k.startsWith('placeholder-variation-label-'))
-              .map((k) => ({ idx: Number.parseInt(k.split('-').pop(), 10), val: (ph[k] || '').trim() }))
-              .filter(({ idx, val }) => Number.isFinite(idx) && val)
-              .sort((a, b) => a.idx - b.idx)
-              .map(({ val }) => val);
-            let urls = [];
-            if (typeof item.variationUrls === 'string' && item.variationUrls.trim()) {
-              const parts = item.variationUrls.split(/\s*,\s*https:\/\//);
-              urls = parts.map((part, idx) => {
-                const trimmed = part.trim();
-                return idx === 0 ? trimmed : `https://${trimmed}`;
-              }).filter((url) => url);
-            }
-            variations = labels.map((lbl, idx) => ({ label: lbl, url: urls[idx] || '' }));
-          }
-          promptMap[item.verb].push({ prompt: item.prompt, assetid: item.assetid, variations });
-        }
-      });
-    }
-    return promptMap;
-  }
-
+  /**
+   * Groups model-picker rows by module (used by tests; marquee uses flat `this.models` from JSON).
+   * @param {Array<object>} data
+   * @returns {Record<string, Array<object>>}
+   */
   createModelMap(data) {
     const modelMap = {};
     if (Array.isArray(data)) {
@@ -627,15 +347,10 @@ export default class UnityWidget {
     return modelMap;
   }
 
-  async updateDropdownForVerb(verb) {
-    if (!this.hasPromptSuggestions) return;
-    await this.ensureSoundModuleLoaded();
-    const dropdown = this.widget.querySelector('#prompt-dropdown');
-    if (!dropdown) return;
-    dropdown.querySelectorAll('.drop-item').forEach((item) => item.remove());
-    const prompts = await this.getPrompt(verb);
-    const limited = this.getLimitedDisplayPrompts(prompts);
-    this.addPromptItemsToDropdown(dropdown, limited, this.workflowCfg.placeholder);
-    this.widgetWrap.dispatchEvent(new CustomEvent('firefly-reinit-action-listeners'));
+  /**
+   * Marquee subclass refreshes prompt suggestions; base is a no-op (style launcher has no suggestions).
+   */
+  async updateDropdownForVerb() {
+    await Promise.resolve();
   }
 }
